Updates backend controllers to be more organized and adds auth error handling.
