+++ backend/routes/userRoutes.js
@@ -1,119 +1,20 @@
 const express = require("express");
 const router = express.Router();
-const User = require("../models/User");
+const usersController = require("../controllers/usersController");
 
 // GET user by uid
-router.get("/:uid", async (req, res) => {
-  try {
-    const user = await User.findOne({ uid: req.params.uid });
-    if (!user) return res.status(404).json({ error: "User not found" });
-    res.json(user);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-});
+router.get("/:uid", usersController.fetchUserByUid);
 
 // GET user by _id
-router.get("/:_id", async (req, res) => {
-  try {
-    const user = await User.findOne({ _id: req.params._id });
-    if (!user) return res.status(404).json({ error: "User not found" });
-    res.json(user);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-});
+router.get("/:_id", usersController.fetchUserById);
 
 // POST create user
-router.post("/", async (req, res) => {
-  try {
-    const { uid, email, firstName, lastName, cart, orders } = req.body;
-
-    if (!uid) {
-      return res.status(400).json({ error: "UID is required" });
-    }
-    if (!email) {
-      return res.status(400).json({ error: "Email is required" });
-    }
-
-    // Check if user already exists
-    const existingUser = await User.findOne({ uid });
-    if (existingUser) {
-      return res.status(409).json({ error: "User already exists" });
-    }
-
-    const newUser = new User({
-      uid,
-      email,
-      firstName: firstName || "Guest",
-      lastName: lastName || "",
-      cart: cart || { cart_items: [], total_items: 0, subtotal: 0 },
-      orders: orders || [],
-    });
-
-    const savedUser = await newUser.save();
-    res.status(201).json(savedUser);
-  } catch (err) {
-    console.error("Error creating user:", err);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-});
+router.post("/", usersController.createUser);
 
 // PUT add to cart
-router.put("/:uid/cart", async (req, res) => {
-  try {
-    const { cart_items, total_items, subtotal } = req.body;
-    const user = await User.findOneAndUpdate(
-      { uid: req.params.uid },
-      {
-        $set: {
-          "cart.cart_items": cart_items,
-          "cart.total_items": total_items,
-          "cart.subtotal": subtotal,
-        },
-      },
-      { new: true }
-    );
-    if (!user) return res.status(404).json({ error: "User not found" });
-    res.json(user);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-});
+router.put("/:uid/cart", usersController.updateUserCart);
 
 // PUT add an order
-router.put("/:uid/orders", async (req, res) => {
-  try {
-    const { orderData } = req.body;
-
-    // Validate order data
-    if (!orderData || typeof orderData !== "object") {
-      return res.status(400).json({ error: "Invalid order data" });
-    }
-
-    // Add new order to the 'orders' array and clear the cart in one query
-    const user = await User.findOneAndUpdate(
-      { uid: req.params.uid },
-      {
-        $push: { orders: orderData },
-        $set: {
-          "cart.cart_items": [],
-          "cart.total_items": 0,
-          "cart.subtotal": 0,
-        },
-      },
-      { new: true }
-    );
-
-    if (!user) {
-      return res.status(404).json({ error: "User not found" });
-    }
-
-    res.json(user); // Return updated user document with new order + cleared cart
-  } catch (err) {
-    console.error("Error creating order:", err);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-});
+router.put("/:uid/orders", usersController.updateUserOrders);
 
 module.exports = router;
