+++ client/.env.development.local.example
@@ -1,4 +1,4 @@
-REACT_APP_BASE_URL=http://localhost:5000
+REACT_APP_BASE_URL=http://localhost:5001
 REACT_APP_FIREBASE_API_KEY=REACT_APP_FIREBASE_API_KEY
 REACT_APP_AUTH_DOMAIN=REACT_APP_AUTH_DOMAIN
 REACT_APP_PROJECT_ID=REACT_APP_PROJECT_ID
