+++ client/.env.production.local.example
@@ -1,1 +1,7 @@
 REACT_APP_BASE_URL=https://stephanies-website-backend.onrender.com
+REACT_APP_FIREBASE_API_KEY=REACT_APP_FIREBASE_API_KEY
+REACT_APP_AUTH_DOMAIN=REACT_APP_AUTH_DOMAIN
+REACT_APP_PROJECT_ID=REACT_APP_PROJECT_ID
+REACT_APP_STORAGE_BUCKET=REACT_APP_STORAGE_BUCKET
+REACT_APP_MESSAGING_SENDER_ID=REACT_APP_MESSAGING_SENDER_ID
+REACT_APP_APP_ID=REACT_APP_APP_ID
