+++ client/src/api/mongoRequests.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import { userModel } from "../Models/User";
 import { _fetchSessionLineItems } from "./stripeRequests";
 
-const baseUrl = process.env.REACT_APP_BASE_URL || "http://localhost:5000";
+const baseUrl = process.env.REACT_APP_BASE_URL || "http://localhost:5001";
 
 /********************************************** User Requests ***********************************************/
 
@@ -12,7 +12,7 @@ const baseUrl = process.env.REACT_APP_BASE_URL || "http://localhost:5000";
  * @returns Returns the newly created user document
  */
 export const _createUser = async (userData) => {
-  const response = await axios.post(`${baseUrl}/api/users`, userData);
+  const response = await axios.post(`${baseUrl}/users`, userData);
   return response.data;
 };
 
@@ -22,7 +22,7 @@ export const _createUser = async (userData) => {
  * @returns Fetches user from mongoDB by uid.
  */
 export const _fetchUserByUid = async (uid) => {
-  const response = await axios.get(`${baseUrl}/api/users/${uid}`);
+  const response = await axios.get(`${baseUrl}/users/${uid}`);
   return response.data;
 };
 
@@ -36,7 +36,7 @@ export const _fetchUserByUid = async (uid) => {
  * @returns {Promise<orderObject[]>} An array of order objects.
  */
 export const _fetchOrdersByUid = async (uid) => {
-  const { data } = await axios.get(`${baseUrl}/api/orders/${uid}`);
+  const { data } = await axios.get(`${baseUrl}/orders/${uid}`);
   return data;
 };
 
@@ -52,7 +52,7 @@ export const _createOrder = async (sessionId, uid, subtotal) => {
   const lineItems = await _fetchSessionLineItems(sessionId);
 
   // Pass in lineItems to create an order object in mongoDB
-  const { data: newOrder } = await axios.post(`${baseUrl}/api/orders`, {
+  const { data: newOrder } = await axios.post(`${baseUrl}/orders`, {
     uid,
     sessionId,
     lineItems,
@@ -71,7 +71,7 @@ export const _createOrder = async (sessionId, uid, subtotal) => {
  * @returns Empties user mongoDB cart
  */
 export const _emptyCart = async (uid) => {
-  const response = await axios.put(`${baseUrl}/api/users/${uid}/cart`, {
+  const response = await axios.put(`${baseUrl}/users/${uid}/cart`, {
     cart_items: [],
     total_items: 0,
     subtotal: 0,
@@ -86,7 +86,7 @@ export const _emptyCart = async (uid) => {
  * @returns Adds products to user's mongoDB cart
  */
 export const _addToCart = async (uid, cart) => {
-  const response = await axios.put(`${baseUrl}/api/users/${uid}/cart`, cart);
+  const response = await axios.put(`${baseUrl}/users/${uid}/cart`, cart);
   return response.data;
 };
 
@@ -97,7 +97,7 @@ export const _addToCart = async (uid, cart) => {
  * @returns Will either decrement or increment the user's product quantity in mongoDB cart
  */
 export const _updateProductQuantity = async (uid, cart) => {
-  const response = await axios.put(`${baseUrl}/api/users/${uid}/cart`, cart);
+  const response = await axios.put(`${baseUrl}/users/${uid}/cart`, cart);
   return response.data;
 };
 
@@ -108,7 +108,7 @@ export const _updateProductQuantity = async (uid, cart) => {
  * @returns Will either decrement or increment the user's product quantity in mongoDB cart
  */
 export const _removeProductFromCart = async (uid, cart) => {
-  const response = await axios.put(`${baseUrl}/api/users/${uid}/cart`, cart);
+  const response = await axios.put(`${baseUrl}/users/${uid}/cart`, cart);
   return response.data;
 };
