+++ client/src/components/AppComponents/AuthModal/AuthModal.scss
@@ -67,3 +67,10 @@ h2 {
   border: 1px solid #cc34ab;
   width: 100%;
 }
+
+.auth-error {
+  color: red;
+  margin-top: 10px;
+  font-size: 0.9rem;
+  text-align: center;
+}
